pyproject.toml
[project]
name = "recommendations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kagglehub>=0.3.13",
    "numpy>=2.3.3",
    "openai>=2.5.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
]
